misc/SpatialEdit/requirements.txt
torch==2.7.1
torchvision==0.22.1
diffusers==0.36.0
transformers==4.57.6
accelerate==1.6.0
peft==0.17.0
einops==0.8.1
safetensors==0.7.0
pillow==11.2.1
pillow-heif==0.21.0
tqdm==4.67.1
loguru==0.7.3
ftfy==6.3.1
opencv-python-headless==4.12.0.88
imageio==2.37.0
imageio-ffmpeg==0.6.0
av==15.0.0
pathvalidate==3.3.1
flask==3.1.1
